fix: fix circular dependency on rules matchers
